GO-6946 Split service.go
